requirements.txt
cryptography
Pillow
numpy